Updated scripts and guide
Turned out zram eventually does have multithreading and you don't really need multiple zram devices. Also turned out disksize of zram is the total uncompressed size of zram, not the amount of actual space it will take up, so if you set zram device disk size to 4 GB, actual size of it when it's full will be less than 2 GB because of compression, so you need to set zram device size to the size of your RAM, it will take up less than half of it when full. I've read it is also safe to make disk size 50% more than the RAM size, didn't test it, maybe will test it later.